Addressed PR comments, restructure the shell handler for Tracing commands
