+++ examples/platform/silabs/shell/BUILD.gn
@@ -23,17 +23,32 @@ if (use_SiWx917) {
   shell_dependency_path = "${chip_root}/examples/platform/silabs/SiWx917"
 }
 
-config("shell-config") {
-  include_dirs = [ "." ]
+config("icd-shell-config") {
+  include_dirs = [ "./icd/" ]
+}
+
+config("tracing-shell-config") {
+  include_dirs = [ "./tracing/" ]
 }
 
 source_set("icd") {
   sources = [
-    "ICDShellCommands.cpp",
-    "ICDShellCommands.h",
+    "./icd/ICDShellCommands.cpp",
+    "./icd/ICDShellCommands.h",
   ]
 
-  public_configs = [ ":shell-config" ]
+  public_configs = [ ":icd-shell-config" ]
+
+  deps = [ "${shell_dependency_path}:matter-shell" ]
+}
+
+source_set("tracing") {
+  sources = [
+    "./tracing/TracingShellCommands.cpp",
+    "./tracing/TracingShellCommands.h",
+  ]
+
+  public_configs = [ ":tracing-shell-config" ]
 
   deps = [ "${shell_dependency_path}:matter-shell" ]
 }
