+++ examples/platform/silabs/shell/icd/ICDShellCommands.cpp
