+++ examples/platform/silabs/shell/icd/ICDShellCommands.h
