+++ examples/platform/silabs/shell/tracing/TracingShellCommands.h
@@ -1,0 +1,23 @@
+/***************************************************************************
+ * @file SilabsTracing.cpp
+ * @brief Instrumenting for matter operation tracing for the Silicon Labs platform.
+ *******************************************************************************
+ * # License
+ * <b>Copyright 2024 Silicon Laboratories Inc. www.silabs.com</b>
+ *******************************************************************************
+ *
+ * The licensor of this software is Silicon Laboratories Inc. Your use of this
+ * software is governed by the terms of Silicon Labs Master Software License
+ * Agreement (MSLA) available at
+ * www.silabs.com/about-us/legal/master-software-license-agreement. This
+ * software is distributed to you in Source Code format and is governed by the
+ * sections of the MSLA applicable to Source Code.
+ *
+ ******************************************************************************/
+#pragma once
+
+namespace TracingCommands {
+
+void RegisterCommands();
+
+} // namespace TracingCommands
